# Improve performance of EvictedHashMap