--- a/opentelemetry/benches/trace.rs
+++ b/opentelemetry/benches/trace.rs
@@ -2,10 +2,25 @@ use criterion::{criterion_group, criterion_main, Criterion};
 use opentelemetry::{
     sdk::trace as sdktrace,
     trace::{Span, Tracer, TracerProvider},
-    Key,
+    Key, KeyValue,
 };
 
 fn criterion_benchmark(c: &mut Criterion) {
+    let mut group = c.benchmark_group("EvictedHashMap");
+    group.bench_function("insert 1", |b| {
+        b.iter(|| insert_keys(sdktrace::EvictedHashMap::new(32, 1), 1))
+    });
+    group.bench_function("insert 5", |b| {
+        b.iter(|| insert_keys(sdktrace::EvictedHashMap::new(32, 5), 5))
+    });
+    group.bench_function("insert 10", |b| {
+        b.iter(|| insert_keys(sdktrace::EvictedHashMap::new(32, 10), 10))
+    });
+    group.bench_function("insert 20", |b| {
+        b.iter(|| insert_keys(sdktrace::EvictedHashMap::new(32, 20), 20))
+    });
+    group.finish();
+
     trace_benchmark_group(c, "start-end-span", |tracer| tracer.start("foo").end());
 
     trace_benchmark_group(c, "start-end-span-4-attrs", |tracer| {
@@ -48,6 +63,35 @@ fn criterion_benchmark(c: &mut Criterion) {
         span.set_attribute(Key::new("key15").f64(123.456));
         span.end();
     });
+}
+
+const MAP_KEYS: [Key; 20] = [
+    Key::from_static_str("key1"),
+    Key::from_static_str("key2"),
+    Key::from_static_str("key3"),
+    Key::from_static_str("key4"),
+    Key::from_static_str("key5"),
+    Key::from_static_str("key6"),
+    Key::from_static_str("key7"),
+    Key::from_static_str("key8"),
+    Key::from_static_str("key9"),
+    Key::from_static_str("key10"),
+    Key::from_static_str("key11"),
+    Key::from_static_str("key12"),
+    Key::from_static_str("key13"),
+    Key::from_static_str("key14"),
+    Key::from_static_str("key15"),
+    Key::from_static_str("key16"),
+    Key::from_static_str("key17"),
+    Key::from_static_str("key18"),
+    Key::from_static_str("key19"),
+    Key::from_static_str("key20"),
+];
+
+fn insert_keys(mut map: sdktrace::EvictedHashMap, n: usize) {
+    for idx in 0..n {
+        map.insert(KeyValue::new(MAP_KEYS[idx].clone(), idx as i64));
+    }
 }
 
 fn trace_benchmark_group<F: Fn(&sdktrace::Tracer)>(c: &mut Criterion, name: &str, f: F) {