A large portion of tracing time is typically spent processing span attributes. This patch improves performance of inserting additional span attributes at creation time by optimizing the backing evicted hash map data structure.

Changes:

* Use `entry` api on insert to avoid hashing keys twice
* Specify an initial attribute capacity when building a span to avoid growing backing hashmap capacity repeatedly
* Avoid allocating and appending extra links vec if unused
* (non-perf) add new type for `Iter` and `IntoIter` to avoid std hashmap in public api

Benchmark changes:

```
EvictedHashMap/insert 1 time:   [210.50 ns 212.56 ns 215.01 ns]
                        change: [-6.6266% -4.9711% -3.3843%] (p = 0.00 < 0.05)
                        Performance has improved.
EvictedHashMap/insert 5 time:   [828.72 ns 836.35 ns 843.99 ns]
                        change: [-21.606% -20.885% -20.169%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild
EvictedHashMap/insert 10
                        time:   [1.5435 us 1.5517 us 1.5604 us]
                        change: [-26.554% -26.207% -25.837%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
EvictedHashMap/insert 20
                        time:   [2.9865 us 2.9960 us 3.0070 us]
                        change: [-28.683% -27.587% -26.707%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe
```